% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.plot.R
\name{cluster.plot}
\alias{cluster.plot}
\title{Visualize Cluster Assignments with Pairwise Scatterplots}
\usage{
cluster.plot(
  db,
  var,
  postprob = 1,
  errorPatterns = ncol(data),
  threshold = 0.6
)
}
\arguments{
\item{db}{A numeric data frame or matrix containing the variables to plot.}

\item{var}{A character vector indicating the names of the ckusters numeric variables in `db. All specified variables must be numeric and positive.}

\item{postprob}{A matrix or dataframe of posterior probabilities where each column represents a cluster.
It is typically obtained from the function \code{assign.cluster()}. If set to 1 (default), it is internally set to a matrix
assigning all observations to a single cluster. They must have the same row number of 'db'}

\item{errorPatterns}{A list of positive numeric vectors that have been identified as being associated with potential unit errors
which will be highlighted in the graphs. If not provided, no special errorPatterns are highlighted (default = ncol(data), used internally)}

\item{threshold}{Minimum probability threshold for cluster assignment (default = 0.6).}
}
\value{
No value returned. The function produces side-effect graphs, which are useful for visually inspecting
the separation between clusters and the reliability of assignments.
}
\description{
This function dynamically generates a set of scatter plots for all pairs of variables, log-transformed, in the provided dataset,
enabling detailed visualisation of cluster assignments.
The points are coloured according to the cluster to which each observation is assigned, based on posterior probabilities.
If the number of variables is large, the number of plots is automatically adjusted to maintain readability,
creating multiple graphs.
Observations are only assigned to a cluster if the a posteriori probability for that cluster is above a specified threshold (default = 0.6).
}
\details{
It is the \strong{user's responsibility} to ensure that input variables are:
\itemize{
  \item Numeric
  \item Strictly positive (since log transformation is applied)
  \item Appropriately scaled (e.g., unit conversion issues should be captured by `errorPatterns`)
}
If these conditions are not met, the function will return informative error messages and halt execution.

Input data is not preprocessed automatically: \strong{preprocessing must be handled by the user}.
}
\examples{
data <- data.frame(
 x1 = c(rlnorm(50, meanlog = 4, sdlog = 0.3), rlnorm(10, meanlog = 4, sdlog = 0.3) * 1000),
 x2 = rlnorm(60, meanlog = 6, sdlog = 0.3)
)

var <- names(data)
# Define two errorPatterns: no shift, and a ×1000 shift in x1 only
errorPatterns <- list(
  c(0, 0),            # Cluster 1: correct units
  c(1000, 0)          # Cluster 2: scale error ×1000
)

results <- assign.cluster(
  db = data,
  var= var,
  errorPatterns = errorPatterns
)

cluster.plot(data, var, postprob = results$postprob, errorPatterns, threshold = 0.6)
}
\seealso{
\code{\link{assign.cluster}}
}
